+++ Engine/Source/Runtime/Engine/Private/UnrealEngine.cpp
@@ -6721,12 +6721,12 @@ float UEngine::GetMaxTickRate(float DeltaTime, bool bAllowFrameRateSmoothing) co
 	return MaxTickRate;
 }
 
-int32 UEngine::GetMaxFPS() const
+float UEngine::GetMaxFPS() const
 {
 	return CVarMaxFPS.GetValueOnAnyThread();
 }
 
-void UEngine::SetMaxFPS(const int32 MaxFPS)
+void UEngine::SetMaxFPS(const float MaxFPS)
 {
 	IConsoleVariable* ConsoleVariable = CVarMaxFPS.AsVariable();
 	ConsoleVariable->Set(MaxFPS);
